Fixed text batch_indices break

+++ CANaRi/test_model/canari.py
@@ -165,7 +165,7 @@ class CANARI(CLIP):
 
         if text is not None:
             if batch_indices is None:
-                batch_indices = torch.arange(len(text), dtype=self.dtype, device=self.device) # Q: should it be dtype=torch.int64 ?
+                batch_indices = torch.arange(len(text), dtype=torch.int64, device=self.device) # Q: should it be dtype=torch.int64 ? > Yes, otherwise breaks
 
             text_features = self.encode_text(text, '{}', batch_indices)
             text_features = text_features / text_features.norm(dim=-1, keepdim=True)
